Add a scheduler size test
Try to determine the size of the scheduler by using padding
of add instructions which all depend on the prior load. These
won't leave the scheduler until the load completes, so the testing
principle works as before, except rather than resources being
freed at retirement, they are freed after execution (so we need
the dependency to avoid that happening).

Reports 64 on my Skylake, which lines up with other tests that
show that SKL actually has a 64-entry scheduler for ALU ops, with
the "97 entry" scheduler actually composed of 64-entry ALU
scheduler and a 33 entry memory operation scheduler.

In the case of non-unified schedulers, we expect this to report
the total size of all the schedulers that can accept add operations,
as long as distribution between the schedulers is even.